# REST APIs Design/openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Title",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    }
  ],
  "paths": {
...
    "/invitation": {
      "post": {
        "operationId": "issue new invitation",
        "requestBody": {
          "description": "issue new invitation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/issueInvitationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "the response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "$ref": "#/components/schemas/invitationModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invitation/:id": {
      "delete": {
        "operationId": "delete an invitation",
        "description": "delete an invitation by its ID",
        "responses": {
          "200": {
            "description": "when the film deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "result": "True"
                }
              }
            }
          },
          "404": {
            "description": "when the film not found",
            "content": {
              "application/json": {
                "example": {
                  "result": "False"
                }
              }
            }
          }
        }
      }
    },
...
    "/films": {
      "get": {
        "operationId": "get films",
        "description": "get films (public films and user films)",
        "responses": {
          "200": {
            "description": "list of films",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/filmsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create new film",
        "requestBody": {
          "description": "the request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createFilmRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "the response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createFilmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/films/:id": {
      "get": {
        "operationId": "get films by Id",
        "responses": {
          "200": {
            "description": "the film",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/filmResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete film",
        "description": "delete film by its ID",
        "responses": {
          "200": {
            "description": "when the film deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "result": "True"
                }
              }
            }
          },
          "404": {
            "description": "when the film not found",
            "content": {
              "application/json": {
                "example": {
                  "result": "False"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "update film",
        "description": "update film by its Id",
        "requestBody": {
          "description": "the updated data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createFilmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "when the film is undated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/filmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/films/media/:id": {
      "delete": {
        "operationId": "delete media",
        "description": "delete media by its ID",
        "responses": {
          "200": {
            "description": "when the media deleted successfully",
            "content": {
              "application/json": {
                "example": {
                  "result": "True"
                }
              }
            }
          },
          "404": {
            "description": "when the media not found",
            "content": {
              "application/json": {
                "example": {
                  "result": "False"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "issueInvitationRequest": {
        "type": "object",
        "properties": {
          "filmId": { "type": "integer" },
          "invitedUserId": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "invitationModel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "filmId": { "type": "integer" },
          "issuedById": { "type": "integer" },
          "invitedUserId": { "type": "integer" },
          "issuer": {
            "$ref": "#/components/schemas/userModel"
          },
          "reviewer": {
            "$ref": "#/components/schemas/userModel"
          },
          "film": {
            "$ref": "#/components/schemas/film"
          },
          "done": {
            "type": "boolean"
          }
        }
      },
...
      "filmResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "title": { "type": "string" },
          "ownerId": { "type": "number" },
          "private": { "type": "boolean" },
          "watchDate": { "$ref": "#/components/schemas/_date" },
          "rating": { "type": "number" },
          "favorite": { "type": "boolean" }
        },
        "example": "{\"id\":0,\"title\":\"Film 1\",\"ownerId\":3,\"private\":true,\"watchDate\":\"2022-12-30T00:00:00.000Z\",\"rating\":2,\"favorite\":false}"
      },
      "filmsResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "$ref": "#/components/schemas/filmResponse"
        },
        "example": "[{\"id\":0,\"title\":\"Film 1\",\"ownerId\":3,\"private\":true,\"watchDate\":\"2022-12-30T00:00:00.000Z\",\"rating\":2,\"favorite\":false},{\"id\":1,\"title\":\"Film 3\",\"ownerId\":3,\"private\":true,\"watchDate\":\"2022-12-13T00:00:00.000Z\",\"rating\":3,\"favorite\":false}]"
      },
      "createFilmRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "isPrivate": { "type": "boolean", "default": true },
          "watchDate": { "$ref": "#/components/schemas/_date" },
          "rating": { "type": "integer", "minimum": 0, "maximum": 10 },
          "favorite": { "type": "boolean", "default": false },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "data": { "type": "string" },
                "name": { "type": "string" }
              }
            }
          }
        }
      },
      "createFilmResponse": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "private": { "type": "boolean", "default": true },
          "watchDate": { "$ref": "#/components/schemas/_date" },
          "rating": { "type": "integer", "minimum": 0, "maximum": 10 },
          "favorite": { "type": "boolean", "default": false },
          "ownerId": { "type": "number" }
        }
      },
...
      "_date": {
        "type": "string",
        "pattern": "^(19|20)[0-9]{2}[- /.](0[1-9]|1[012])[- /.](0[1-9]|[12][0-9]|3[01])$"
      },
      "userModel": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6, "maxLength": 20 }
        }
      },
      "film": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "ownerId": { "type": "integer" },
          "private": { "type": "boolean", "default": true },
          "watchDate": { "$ref": "#/components/schemas/_date" },
          "rating": { "type": "integer", "minimum": 0, "maximum": 10 },
          "favorite": { "type": "boolean", "default": false },
          "medias": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/media"
            }
          }
        }
      },
      "media": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "type": { "type": "string" },
          "name": { "type": "string" },
          "filmId": { "type": "integer" }
        }
      },
...
    }
  }
}